+++ tests/Unit/Private/Assert-AzureConnection.tests.ps1
@@ -105,9 +105,6 @@ Describe 'Assert-AzureConnection' {
                     }
                 } | Should -Throw
             }
-
         }
     }
-
-
 }